LegalDoc_GenAI/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.5
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.36.0
google-generativeai==0.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.27
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.2
langsmith==0.1.147
lxml==6.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.2.0
networkx==3.4.2
numpy==1.26.4
orjson==3.11.3
packaging==24.2
pandas==2.3.2
pillow==10.4.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.36.0
tenacity==8.5.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.66.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1